requirements.txt
numpy
opencv-python
tensorflow
Pillow
psycopg2-binary
requests